[EDITED] fully implemented InputManager [REFACTORED] got rid of unnecessary code in demo.cpp now that InputManager is fully implemented [FIXED] got rid of warnings in eases.h by adding "f" after each number so the compiler doesn't mistake them for doubles; got rid of warnings in InputManager

+++ EGP Graphics FW Class Base/source/egpfwDemo/utils/SaveManager.cpp
@@ -68,7 +68,7 @@ void SaveManager::loadData(int dataFileSelected)
 			}
 			else if (type == "float")
 			{
-				m_mapFloat[name] = atof(data.c_str());
+				m_mapFloat[name] = std::stof(data.c_str());
 				//m_mapFloat.insert(std::pair<std::string, float>(name, static_cast<float>(atof(data.c_str()))));
 			}
 			else if (type == "bool")
